# openapi/attempts.yaml
openapi: 3.0.4
info:
  title: N2 Daily Mock Exam — Attempts API
  version: 0.1.0
  description: |
    HTTP contract for the /api/attempts family of endpoints in the N2 Daily Mock
    Exam webapp (Next.js App Router + Supabase). This spec is the canonical,
    machine-readable description of the request/response shape that
    `app/api/attempts/route.ts` and `app/api/attempts/[date]/route.ts`
    implement. Webapp consumers (including the /today submit handler, the
    /result/[date] result page loader, and any future client SDK or
    TypeScript codegen target) should derive their types from this spec.

    Authentication: cookie-based Supabase session. Endpoints read
    `supabase.auth.getSession()` via `createServerSupabase()`; an absent
    session returns 401. There is no Bearer-token path; magic-link flow
    (see /auth/callback) is the entry point.

    Status-code semantics:
      * 200 — success (including the "no mock yet" path on GET — the body
        returns { mock: null, attempts: [], score: 0, total: 0,
        total_seconds: 0 }).
      * 400 — client-side validation failure (invalid JSON, missing
        required fields, malformed path/date).
      * 401 — no Supabase session cookie.
      * 500 — DB insert/query failure.
  contact:
    name: n2-build loop
    url: ~/ODIS/shared/loops/n2-build/
servers:
  - url: http://localhost:3000
    description: Local Next.js dev server
  - url: https://n2.example.com
    description: Production (placeholder — replace with deployed origin)

tags:
  - name: attempts
    description: |
      Per-attempt persistence + retrieval. The submit path (POST) writes
      per-question rows into n2_attempts; the read path (GET) returns the
      same rows enriched with question context from n2_questions, ready for
      the result page to render.

paths:
  /api/attempts:
    post:
      tags: [attempts]
      operationId: submitAttempt
      summary: Submit answers for today's mock
      description: |
        Inserts one row per answered question into `n2_attempts`. The
        question-granular shape (rather than one row per mock) is what the
        cp6 result page expects — each row carries its own `time_seconds`,
        `user_answer`, and `correct` flag.

        The handler derives `total_ms` as the sum of `timings` values when
        the client omits it.

        M4: `challenge_id` is inserted as `null` (migration 004 — the
        `n2_mocks` PK is `DATE`, not UUID). `question_id` is the join key
        used by GET /api/attempts/{date} to scope attempts to a mock.

        M6: `correct_map` from the client carries
        `Question.correct_answer` per question. When supplied, `correct`
        is populated at insert time (boolean). Omitting `correct_map` is
        backwards-compatible — `correct` stays `null` and is recomputed
        on read.

        Auth: requires a Supabase session cookie. Returns 401 otherwise.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAttemptRequest'
            examples:
              minimal:
                summary: Minimal valid payload
                value:
                  date: '2026-07-13'
                  answers:
                    'q-uuid-1': 'b'
                    'q-uuid-2': 'c'
                    'q-uuid-3': 'a'
                    'q-uuid-4': 'd'
                    'q-uuid-5': 'b'
                  timings:
                    'q-uuid-1': 12000
                    'q-uuid-2': 18000
                    'q-uuid-3': 9000
                    'q-uuid-4': 22000
                    'q-uuid-5': 15000
              withCorrectMap:
                summary: With per-question correct flags pre-computed client-side
                value:
                  date: '2026-07-13'
                  answers:
                    'q-uuid-1': 'b'
                    'q-uuid-2': 'c'
                  timings:
                    'q-uuid-1': 12000
                    'q-uuid-2': 18000
                  correct_map:
                    'q-uuid-1': 'b'
                    'q-uuid-2': 'c'
                  total_ms: 30000
      responses:
        '200':
          description: Attempt rows persisted; returns the inserted row IDs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitAttemptResponse'
              example:
                attempt_ids:
                  - '9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d'
                  - '1b9d6bcd-bbfd-4b2d-9b5d-ab8dfbbd4bed'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/attempts/{date}:
    get:
      tags: [attempts]
      operationId: getAttemptByDate
      summary: Read the authenticated user's attempt for a given daily mock
      description: |
        Returns the user's attempt rows for the daily mock published on
        `{date}` (YYYY-MM-DD), enriched with the full question metadata
        join from `n2_questions`. Drives the `/result/[date]` result
        page.

        Behaviour:
          1. Look up `n2_mocks` for `{date}`. Missing mock → return the
             "no mock yet" stub (200 with empty arrays).
          2. Pull the user's `n2_attempts` rows whose `question_id` is in
             `mock.question_ids`. `challenge_id` is ignored (nullable per
             migration 004).
          3. Pull the matching `n2_questions` rows (one round-trip).
          4. Enrich each attempt with `is_correct` (recomputed from
             `question.correct_answer === user_answer`) and the full
             question row.

        Returns 200 in all success cases (including no-mock + no-attempt
        paths). Only 400, 401, 500 carry an error.
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
            example: '2026-07-13'
          description: Daily mock date (YYYY-MM-DD).
      responses:
        '200':
          description: |
            Attempt payload. The `mock` field is `null` when no mock has
            been published for `{date}` (the result page renders a
            friendly empty state in that case).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAttemptResponse'
              examples:
                withAttempts:
                  summary: User completed today's mock
                  value:
                    mock:
                      id: null
                      date: '2026-07-13'
                      questions:
                        - 'q-uuid-1'
                        - 'q-uuid-2'
                        - 'q-uuid-3'
                        - 'q-uuid-4'
                        - 'q-uuid-5'
                    attempts:
                      - question_id: 'q-uuid-1'
                        user_answer: 'b'
                        time_seconds: 12
                        is_correct: true
                        question:
                          id: 'q-uuid-1'
                          category: grammar_form
                          tags: ['particle', 'past-tense']
                          difficulty: 3
                          prompt: |
                            昨日 本を___。
                          options:
                            - id: a
                              text: '読みました'
                            - id: b
                              text: '読みました'  # note: distinct from a in real data
                            - id: c
                              text: '読む'
                            - id: d
                              text: '読んで'
                          correct_answer: 'b'
                          explanation: |
                            The past-tense polite form 読みました fits the
                            sentence-ending particle と and temporal clause
                            昨日.
                          source_id: 'shin-n2-2024-q42'
                          audit:
                            score: 8.4
                            scores_by_criterion:
                              japanese_accuracy: 9.0
                              n2_alignment: 8.0
                              clarity: 8.5
                              distractor_quality: 8.0
                              format_compliance: 8.5
                            reviewer_model: 'claude-sonnet-4-5'
                            reviewed_at: '2026-07-13T01:05:00+00:00'
                    score: 4
                    total: 5
                    total_seconds: 96
                noMock:
                  summary: No mock published for this date yet
                  value:
                    mock: null
                    attempts: []
                    score: 0
                    total: 0
                    total_seconds: 0
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    AnswerValue:
      type: string
      enum: [a, b, c, d]
      description: Multiple-choice option selector.
    AnswerMap:
      type: object
      description: |
        Map from question UUID (`n2_questions.id`) to the user's chosen
        answer. Keys must be the UUID strings — they are passed through
        verbatim into the `.eq()`/`.in()` query filter.
      additionalProperties:
        $ref: '#/components/schemas/AnswerValue'
    TimingMap:
      type: object
      description: |
        Map from question UUID to per-question timing in **milliseconds**
        (the POST handler divides by 1000 before storing in `time_seconds`).
      additionalProperties:
        type: integer
        minimum: 0
    CorrectMap:
      type: object
      description: |
        Optional map from question UUID to `Question.correct_answer`. When
        supplied, the handler populates `n2_attempts.correct` at insert
        time (boolean). Omitting the field leaves `correct` as `null` and
        the value is recomputed on read.
      additionalProperties:
        $ref: '#/components/schemas/AnswerValue'
    SubmitAttemptRequest:
      type: object
      required: [date, answers]
      additionalProperties: false
      properties:
        date:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: |
            Daily mock date (must match `n2_mocks.date` — the table PK is
            DATE, not UUID). Validated by the handler with the same regex.
        answers:
          $ref: '#/components/schemas/AnswerMap'
        timings:
          $ref: '#/components/schemas/TimingMap'
        correct_map:
          $ref: '#/components/schemas/CorrectMap'
        total_ms:
          type: integer
          minimum: 0
          description: |
            Optional total elapsed time in milliseconds. If omitted, the
            handler computes the sum of `timings[*]` automatically.
    SubmitAttemptResponse:
      type: object
      required: [attempt_ids]
      additionalProperties: false
      properties:
        attempt_ids:
          type: array
          description: |
            One UUID per answered question (one row per question inserted
            into `n2_attempts`).
          items:
            type: string
            format: uuid
    Question:
      type: object
      description: |
        Canonical question row from `n2_questions` (mirrors the JSON
        Schema at `~/ODIS/shared/loops/n2-build/question.schema.json`).
        M4: `source_id` (post-migration-007 rename); the cp6 result
        page reads this field.
      required: [id, category, options, correct_answer]
      properties:
        id:
          type: string
          format: uuid
        category:
          type: string
          enum: [grammar_form, reading_short, vocab_context]
          description: |
            Launch-categories only. `listening_*` enums are deferred per
            locked goal.md OQ-4 (post-traction).
        tags:
          type: array
          items:
            type: string
        difficulty:
          type: integer
          minimum: 1
          maximum: 5
        prompt:
          type: string
        options:
          type: array
          minItems: 2
          maxItems: 6
          items:
            type: object
            required: [id, text]
            properties:
              id:
                type: string
                enum: [a, b, c, d]
              text:
                type: string
        correct_answer:
          $ref: '#/components/schemas/AnswerValue'
        explanation:
          type: string
        source_id:
          type: string
          description: |
            External/reference source identifier. See migration 007 for
            the rename from `source` → `source_id`.
        audit:
          type: object
          properties:
            score:
              type: number
              minimum: 0
              maximum: 10
            scores_by_criterion:
              type: object
              properties:
                japanese_accuracy:
                  type: number
                n2_alignment:
                  type: number
                clarity:
                  type: number
                distractor_quality:
                  type: number
                format_compliance:
                  type: number
            reviewer_model:
              type: string
            reviewed_at:
              type: string
              format: date-time
    Mock:
      type: object
      required: [date, questions]
      properties:
        id:
          type: string
          format: uuid
          nullable: true
          description: |
            Always `null` in the GET response — `n2_mocks` has a DATE PK,
            not a UUID. The field is reserved for future schema evolution.
        date:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        questions:
          type: array
          description: |
            Ordered list of question UUIDs in `n2_mocks.question_ids`
            (migration 008 column name; NOT `questions`). Preserves
            canonical order on the result page.
          items:
            type: string
            format: uuid
    Attempt:
      type: object
      required: [question_id, user_answer, time_seconds, is_correct]
      properties:
        question_id:
          type: string
          format: uuid
        user_answer:
          $ref: '#/components/schemas/AnswerValue'
        time_seconds:
          type: integer
          minimum: 0
        is_correct:
          type: boolean
        question:
          $ref: '#/components/schemas/Question'
    GetAttemptResponse:
      type: object
      required: [mock, attempts, score, total, total_seconds]
      properties:
        mock:
          $ref: '#/components/schemas/Mock'
        attempts:
          type: array
          items:
            $ref: '#/components/schemas/Attempt'
        score:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
        total_seconds:
          type: integer
          minimum: 0
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Machine-readable error code.
        detail:
          type: string
          description: Optional implementation detail (only on 500s).

  responses:
    BadRequest:
      description: Client-side validation failure (invalid JSON, missing field, malformed date).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'invalid JSON'
    Unauthorized:
      description: No Supabase session cookie present.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: unauthorized
    InternalServerError:
      description: DB insert/query failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'insert failed'
            detail: 'permission denied for table n2_attempts'

security: []  # Endpoints are session-cookie-gated; no global Bearer scheme declared.